Moved auth code to get a user from auth header into its own function getUserFromToken(req); fixed some formatting in routes/index.js; added env variable FOLDER_PATH to README.md env vars section.
