+++ routes/index.js
@@ -5,6 +5,7 @@
  */
 import express from 'express';
 import AppController from '../controllers/AppController';
+import AuthController from '../controllers/AuthController';
 import UsersController from '../controllers/UsersController';
 import FilesController from '../controllers/FilesController';
 
@@ -18,6 +19,14 @@ router.get('/status', AppController.getStatus);
 // GET /stats
 router.get('/stats', AppController.getStats);
 
+// ------------- FilesController Endpoints ------------- \\
+
+// GET /files/:id
+router.get('/files/:id', FilesController.getShow);
+
+// GET /files
+router.get('/files', FilesController.getIndex);
+
 // POST /files
 router.post('/files', FilesController.postUpload);
 
@@ -29,11 +38,6 @@ router.post('/files', FilesController.postUpload);
 // POST /users
 router.post('/users', UsersController.postNew);
 
-// GET /files/:id
-router.get('/files/:id', FilesController.getShow);
-
-// GET /files
-router.get('/files', FilesController.getIndex);
 // ------------- AuthController Endpoints ------------- \\
 
 // GET /connect
